+++ routes/upload.js
@@ -1,7 +1,7 @@
 var utility = require('../utility');
 var domain = require('domain');
+var queueReq = utility.queueReq;
 var multiparty = require('multiparty');
-var db = require('../db');
 var express = require('express');
 var upload = express.Router();
 var Recaptcha = require('recaptcha').Recaptcha;
@@ -16,58 +16,67 @@ upload.get("/", function(req, res) {
 upload.post("/", function(req, res) {
     if (req.session.captcha_verified || process.env.NODE_ENV === "test") {
         req.session.captcha_verified = false; //Set back to false
+        var form = new multiparty.Form();
+        var match_id;
+        var parser;
         var d = domain.create();
-        d.on('error', function() {
-            if (!res.headerSent) {
-                res.render("upload", {
-                    error: "Couldn't parse replay"
-                });
+        d.on('error', function(err) {
+            console.log(err.stack);
+            //user aborted upload
+            if (parser) {
+                parser.kill();
             }
         });
         d.run(function() {
-            var parser = utility.runParse(function(err, output) {
-                if (err) {
-                    throw err;
-                }
-                var match_id = output.match_id;
-                console.log("getting upload data from api");
-                var container = utility.generateJob("api_details", {
-                    match_id: match_id
-                });
-                utility.getData(container.url, function(err, data) {
-                    if (err) {
-                        throw err;
-                    }
-                    var match = data.result;
-                    match.parsed_data = output;
-                    match.parse_status = 2;
-                    match.upload = true;
-                    db.matches.update({
-                        match_id: match_id
-                    }, {
-                        $set: match
-                    }, {
-                        upsert: true
-                    }, function(err) {
-                        if (err) {
-                            throw err;
-                        }
-                        res.redirect("/matches/" + match_id);
-                    });
-                });
+            form.on('error', function(err) {
+                throw err;
             });
-            var form = new multiparty.Form();
+            form.on('field', function(name, value) {
+                console.log('got field named ' + name);
+                //queue for api details, redirect
+                match_id = Number(value);
+                queueReq("api_details", {
+                    match_id: match_id,
+                    priority: "high"
+                }, close);
+            });
             form.on('part', function(part) {
                 if (part.filename) {
+                    console.log('got file named ' + part.name);
+                    //parse to determine match_id, queue for api details, redirect
+                    parser = utility.runParse(function(err, output) {
+                        if (err) {
+                            return close(err);
+                        }
+                        match_id = output.match_id;
+                        queueReq("api_details", {
+                            match_id: match_id,
+                            upload: true,
+                            parsed_data: output,
+                            priority: "high"
+                        }, close);
+                    });
                     part.pipe(parser.stdin);
                 }
             });
-            form.on('error', function(err) {
-                parser.kill();
-                throw err;
-            });
             form.parse(req);
         });
+    }
+
+    function close(err, job) {
+        if (err) {
+            return res.render("upload", {
+                error: err
+            });
+        }
+        else if (job) {
+            job.on('complete', function() {
+                res.redirect("/matches/" + match_id);
+            });
+        }
+        else {
+            res.redirect("/matches/" + match_id);
+        }
     }
 });
 module.exports = upload;
